Refactoring for modularity (untested)
